number of available pages coming back doesnt make sense , investigating. might be an issue with my bootleg int printer

+++ kernel/src/include/uart.h
@@ -8,6 +8,6 @@ void init_serial();
 int is_transmit_empty();
 void write_serial(char a);
 void write_string_serial(const char *str);
-void write_int_serial(int num);
+void write_int_serial(uint64 num);
 void bootleg_panic(const char *str);
 #endif //KERNEL_UART_H
